Major code quality improvements
It took me around 3 hours chasing every integer conversion.

Instead of using signed types in our code to match libav, I've decided
to start casting after some sanity checks.

+++ src/report.c
@@ -174,10 +174,9 @@ anu_report anu_generate_report (anu_file_q *files, uint64_t *hashes,
     /* Important to zero-initialize */
     anu_dupe_group segment_results = {0};
 
-    for (int seg = 0; seg < config->segments; seg++) {
+    for (size_t seg = 0; seg < config->segments; seg++) {
 
-      uint64_t current_hash =
-          hashes[(i * (size_t) config->segments) + (size_t) seg];
+      uint64_t current_hash = hashes[(i * config->segments) + seg];
       bk_tree_search(tree->root, current_hash, config->threshold,
                      &segment_results);
     }
