+++ src/tree.h
@@ -7,8 +7,6 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include "util.h"
-
 typedef struct anu_dupe_group {
   uint64_t files[65];
   size_t file_count;
@@ -32,7 +30,7 @@ typedef struct bk_tree {
 bk_node *bk_tree_node_new(uint64_t hash, uint64_t file_id);
 
 /* Search for nodes with hashes with distance less than tolerance */
-void bk_tree_search(bk_node *node, uint64_t hash, i32 tolerance,
+void bk_tree_search(bk_node *node, uint64_t hash, size_t tolerance,
                     anu_dupe_group *groups_out);
 
 /* Insert hash into tree */
